MLLM backend: refactor by moving TokenSet to annif.lexical.tokenset module
