+++ annif/backend/mllm.py
@@ -16,6 +16,7 @@ from sklearn.tree import DecisionTreeClassifier
 import annif.util
 from annif.exception import NotInitializedException
 from annif.suggestion import VectorSuggestionResult
+from annif.lexical.tokenset import TokenSet, TokenSetIndex
 from . import backend
 from . import hyperopt
 
@@ -33,79 +34,6 @@ Feature = IntEnum(
     'first_occ last_occ spread doc_length ' +
     'broader narrower related',
     start=0)
-
-
-class TokenSet:
-    """Represents a set of tokens (expressed as integer token IDs) that can
-    be matched with another set of tokens. A TokenSet can optionally
-    be associated with a subject from the vocabulary."""
-
-    def __init__(self, tokens, subject_id=None, is_pref=False):
-        self._tokens = set(tokens)
-        self.subject_id = subject_id
-        self.is_pref = is_pref
-
-    def __len__(self):
-        return len(self._tokens)
-
-    def __iter__(self):
-        return iter(self._tokens)
-
-    def contains(self, other):
-        """Returns True iff the tokens in the other TokenSet are all
-        included within this TokenSet."""
-
-        return other._tokens.issubset(self._tokens)
-
-    def sample(self):
-        """Return an arbitrary token from this TokenSet, or None if empty"""
-        try:
-            return next(iter(self._tokens))
-        except StopIteration:
-            return None
-
-
-class TokenSetIndex:
-    """A searchable index of TokenSets (representing vocabulary terms)"""
-
-    def __init__(self):
-        self._index = collections.defaultdict(set)
-
-    def __len__(self):
-        return len(self._index)
-
-    def add(self, tset):
-        """Add a TokenSet into this index"""
-        token = tset.sample()
-        if token is not None:
-            self._index[token].add(tset)
-
-    def search(self, tset):
-        """Return the TokenSets that are contained in the given TokenSet.
-        The matches are returned as a list of (TokenSet, ambiguity) pairs
-        where ambiguity is an integer indicating the number of other TokenSets
-        that also match the same tokens."""
-
-        subj_tsets = {}
-        subj_ambiguity = collections.Counter()
-
-        for token in tset:
-            for ts in self._index[token]:
-                if not tset.contains(ts):
-                    continue
-                if ts.subject_id not in subj_tsets or \
-                   not subj_tsets[ts.subject_id].is_pref:
-                    subj_tsets[ts.subject_id] = ts
-
-        for ts in subj_tsets.values():
-            for other in subj_tsets.values():
-                if ts == other:
-                    continue
-                if other.contains(ts):
-                    subj_ambiguity.update([ts.subject_id])
-
-        return [(ts, subj_ambiguity[ts.subject_id])
-                for uri, ts in subj_tsets.items()]
 
 
 class MLLMModel:
